% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvir-package.R
\docType{package}
\name{curvir-package}
\alias{curvir-package}
\alias{_PACKAGE}
\title{curvir: Specify Reserve Demand Curves}
\description{
Automatic specification and estimation of reserve demand curves for central bank operations. The package can help to choose the best demand curve and identify additional explanatory variables. Various plot and predict options are included. For more details, see Chen et al. (2023) \url{https://www.imf.org/en/Publications/WP/Issues/2023/09/01/Modeling-the-Reserve-Demand-to-Facilitate-Central-Bank-Operations-538754}.
}
\author{
\strong{Maintainer}: Nikolaos Kourentzes \email{nikolaos@kourentzes.com}

Authors:
\itemize{
  \item Zhuohui Chen \email{zchen4@imf.org}
  \item Romain Veyrune
}

}
\keyword{internal}
\keyword{package}
